+++ Runtime/Screens/IDialogue.cs
@@ -1,6 +1,6 @@
 ﻿namespace Samurai.Showcase.Runtime.Screens
 {
-    public interface IDialogue : IScreen
+    internal interface IDialogue
     {
     }
 }