+++ Runtime/Screens/IScreen.cs
@@ -1,9 +1,0 @@
-﻿namespace Samurai.Showcase.Runtime.Screens
-{
-    public interface IScreen
-    {
-        bool IsActive { get; }
-        void Show();
-        void Hide();
-    }
-}