+++ Runtime/Screens/IWindow.cs
@@ -1,6 +1,6 @@
 ﻿namespace Samurai.Showcase.Runtime.Screens
 {
-    public interface IPanel : IScreen
+    internal interface IWindow
     {
     }
 }