+++ Runtime/Screens/Window.cs
@@ -4,7 +4,7 @@
     /// Screen type that can have multiple instances active in the same layer at the same time.
     /// </summary>
     /// <typeparam name="TData">Type of initialization parameters.</typeparam>
-    public abstract class Panel<TData> : Screen<TData>, IPanel
+    public abstract class Window<TData> : Screen<TData>, IWindow
     {
     }
 }